requirements.txt
torch
transformers
openai-whisper
Pillow
requests